requirements.txt

python-chess
requests
gtts
pillow
cairosvg
stockfish
opencv-python
moviepy
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
